# docs/swagger.yaml
basePath: /api
definitions:
  dto.AssignPermissionRequest:
    properties:
      permission_name:
        type: string
      role_name:
        type: string
    required:
    - permission_name
    - role_name
    type: object
  dto.AssignRoleRequest:
    properties:
      role:
        type: string
      user_id:
        type: string
    required:
    - role
    - user_id
    type: object
  dto.CreatePermissionRequest:
    properties:
      name:
        type: string
    required:
    - name
    type: object
  dto.CreateRoleRequest:
    properties:
      name:
        type: string
    required:
    - name
    type: object
  dto.ForgotPasswordRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  dto.ResetPasswordRequest:
    properties:
      code:
        type: string
      email:
        type: string
      new_password:
        minLength: 6
        type: string
    required:
    - code
    - email
    - new_password
    type: object
  dto.UserLoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  dto.UserRegisterRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - name
    - password
    type: object
  dto.UserResponse:
    properties:
      email:
        type: string
      id:
        type: string
      name:
        type: string
      token:
        type: string
    type: object
  dto.VerifyEmailRequest:
    properties:
      code:
        type: string
      email:
        type: string
    required:
    - code
    - email
    type: object
  utils.Pagination:
    properties:
      current_page:
        type: integer
      from:
        type: integer
      has_more_pages:
        type: boolean
      last_page:
        type: integer
      per_page:
        type: integer
      to:
        type: integer
      total:
        type: integer
    type: object
  utils.PaginationResult:
    properties:
      items: {}
      pagination:
        $ref: '#/definitions/utils.Pagination'
    type: object
  utils.Response:
    properties:
      data: {}
      errors: {}
      message:
        type: string
      meta: {}
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is a sample server for Golang Backend.
  title: Golang Backend API
  version: "1.0"
paths:
  /admin/assign-permission:
    post:
      consumes:
      - application/json
      description: Assign a permission to a role (Admin only)
      parameters:
      - description: Assignment Data
        in: body
        name: assignment
        required: true
        schema:
          $ref: '#/definitions/dto.AssignPermissionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Assign a permission to a role
      tags:
      - Roles
  /admin/assign-role:
    post:
      consumes:
      - application/json
      description: Assign a role to a user (Admin only)
      parameters:
      - description: Assignment Data
        in: body
        name: assignment
        required: true
        schema:
          $ref: '#/definitions/dto.AssignRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Assign a role to a user
      tags:
      - Roles
  /admin/permissions:
    post:
      consumes:
      - application/json
      description: Create a new permission (Admin only)
      parameters:
      - description: Permission Data
        in: body
        name: permission
        required: true
        schema:
          $ref: '#/definitions/dto.CreatePermissionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Create a new permission
      tags:
      - Roles
  /admin/roles:
    post:
      consumes:
      - application/json
      description: Create a new role (Admin only)
      parameters:
      - description: Role Data
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/dto.CreateRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Create a new role
      tags:
      - Roles
  /forgot-password:
    post:
      consumes:
      - application/json
      description: Send reset password code to email
      parameters:
      - description: Email
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Forgot Password
      tags:
      - auth
  /login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return JWT token
      parameters:
      - description: Login Credentials
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.UserLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  properties:
                    token:
                      type: string
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Login User
      tags:
      - auth
  /me:
    get:
      consumes:
      - application/json
      description: Get details of the currently logged-in user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Get Current User
      tags:
      - user
  /register:
    post:
      consumes:
      - application/json
      description: Create a new user account
      parameters:
      - description: Register Data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.UserRegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Register User
      tags:
      - auth
  /reset-password:
    post:
      consumes:
      - application/json
      description: Reset password with code
      parameters:
      - description: Reset Data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Reset Password
      tags:
      - auth
  /users:
    get:
      consumes:
      - application/json
      description: Get paginated list of users with search and filter
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 15
        description: Items per page
        in: query
        name: per_page
        type: integer
      - description: Search term
        in: query
        name: search
        type: string
      - description: Filter by verification status
        in: query
        name: is_verified
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.PaginationResult'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Get Users
      tags:
      - user
  /verify-email:
    post:
      consumes:
      - application/json
      description: Verify user email with code
      parameters:
      - description: Verification Code
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.VerifyEmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Verify Email
      tags:
      - auth
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"